added "gen hack" to generate data with hacks

+++ hacks/overflow.cpp
@@ -1,7 +1,8 @@
+// WA 解法：用 int 存储 a+b，遇到大数会溢出
 #include <bits/stdc++.h>
 using namespace std;
 int main(){
-    long long a,b;
+    int a,b;
     cin>>a>>b;
     cout<<a+b<<endl;
     return 0;
